Fix compiler warning.

+++ QGL_Content/qgl_content_dict_entry_buffer.cpp
@@ -3,10 +3,10 @@
 using namespace qgl::content;
 
 CONTENT_DICTIONARY_ENTRY_BUFFER::CONTENT_DICTIONARY_ENTRY_BUFFER() :
-   m_offset(-1),
+   m_offset(static_cast<uint64_t>(-1)),
    m_flags(DEFAULT_FLAGS),
-   m_size(0),
-   m_reserved1(0),
+   m_size(static_cast<uint64_t>(0)),
+   m_reserved1(static_cast<uint64_t>(0)),
    m_info()
 {
 }
@@ -17,7 +17,7 @@ CONTENT_DICTIONARY_ENTRY_BUFFER::CONTENT_DICTIONARY_ENTRY_BUFFER(
    size_t entryOffset) :
    m_offset(entryOffset),
    m_flags(DEFAULT_FLAGS),
-   m_reserved1(0),
+   m_reserved1(static_cast<uint64_t>(0)),
    m_size(entrySize),
    m_info(entryInfo)
 {
